ref: Move limiter instance to separate package
